Update index.md [Private account Git hub]
Hi there, I have taken up the Edx course on Duckiebot. In the course we are instructed to create a private key for GitHub but could not find the same in official documentation. So thought of adding it here as most learns are new to these accounts so will come in handy to them.